Add Javadoc to non-private classes/methods
