+++ src/main/java/duke/task/ToDo.java
@@ -1,11 +1,11 @@
 package duke.task;
 
-public class ToDos extends Task{
+public class ToDo extends Task{
     
-    public ToDos(String description) {
+    public ToDo(String description) {
         super(description);
     }
-
+    
     @Override
     public String toString() {
         return "[T]" + super.toString() ;
